refactor: StatusCode 클래스의 패키지 이동으로 인한 참조 수정
