+++ src/main/java/team05/integrated_feed_backend/exception/custom/DuplicateResourceException.java
@@ -1,7 +1,7 @@
 package team05.integrated_feed_backend.exception.custom;
 
 import lombok.Getter;
-import team05.integrated_feed_backend.exception.code.StatusCode;
+import team05.integrated_feed_backend.common.code.StatusCode;
 
 /**
  * 생성하고자 요청하는 데이터가 이미 있는 경우
