+++ src/main/java/team05/integrated_feed_backend/exception/custom/ForbiddenException.java
@@ -1,7 +1,7 @@
 package team05.integrated_feed_backend.exception.custom;
 
 import lombok.Getter;
-import team05.integrated_feed_backend.exception.code.StatusCode;
+import team05.integrated_feed_backend.common.code.StatusCode;
 
 /**
  * 권한이 없는 곳에 접근하고자 하는 경우
